Fix window transformation related issue
In some cases, for example, a window containing a SurfaceView,
when pressing switch button, the window animation of SurfaceView
and it's attached window are not sync, which leads animation not
behaving correctly.

To avoid this, apply the translation that applies the position
of the window before the app transformation.

Google issue:
https://code.google.com/p/android/issues/detail?id=225134

Change-Id: Ie8ddb875e895942c7f28e6fc83c4f92568512efb